Fix bug: new msg handling blocked node shutdown
This fix will add a flag that indicates the p2p server will be
closed. Once it's set to true, new msg will not be handled
so that new peers will not be added and protocals will not be
initiated any more and the server could be shut down with
no blocking events.